fix project name
